requirements.txt
cython
pillow>=7.0.0
numpy>=1.18.1
opencv-python>=4.1.2
torch>=1.9.0 --extra-index-url https://download.pytorch.org/whl/cpu
torch>=1.9.0 --extra-index-url https://download.pytorch.org/whl/cu102
torchvision>=0.10.0 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision>=0.10.0 --extra-index-url https://download.pytorch.org/whl/cu102
pytest==7.1.3
tqdm==4.64.1
scipy>=1.7.3
matplotlib>=3.4.3
mock==4.0.3